+++ include/docker.bash
@@ -4,7 +4,7 @@ docker-check-boot2docker() {
     if [[ "$missing" ]]; then
         echo "[ERROR] boot2docker command not found, please install by:" | red
         echo "  brew install boot2docker" | blue
-        exit 127
+        _exit 127
     fi
 
     : << "UNTIL-BOOT2DOCKER-CLI-366-GET-MERGED"
@@ -14,12 +14,12 @@ docker-check-boot2docker() {
         else
             echo "[ERROR] boot2docker shell env is not set correctly, please run:" | red
             echo ' eval "$(boot2docker shellinit)"' | blue
-            exit 125
+            _exit 125
         fi
     else
         echo "[ERROR] boot2docker is not running, please start by:" | red
         echo "  boot2docker start" | blue
-        exit 126
+        _exit 126
     fi
 UNTIL-BOOT2DOCKER-CLI-366-GET-MERGED
     if [[ "$(boot2docker status)" == "running" ]]; then
@@ -28,12 +28,12 @@ UNTIL-BOOT2DOCKER-CLI-366-GET-MERGED
         else
             echo "[ERROR] boot2docker shell env is not set correctly, please run:" | red
             echo ' eval "$(boot2docker shellinit)"' | blue
-            exit 125
+            _exit 125
         fi
     else
         echo "[ERROR] boot2docker is not running, please start by:" | red
         echo "  boot2docker start" | blue
-        exit 126
+        _exit 126
     fi
 
     debug "TODO: check for version and instruction for update ..."
@@ -48,7 +48,7 @@ UNTIL-BOOT2DOCKER-CLI-366-GET-MERGED
         localDate=$(date -u +%Y-%m-%d\ %H:%M)
         if [[ "$localDate" != "$b2dDate" ]];then
             echo "Couldnt correct date in boot2docker, giving up" |red
-            exit 2
+            _exit 2
         else
              info "boot2docker date settings: OK" | green
         fi
@@ -78,30 +78,30 @@ docker-check-client-version() {
     docker --version &> /dev/null || local missing=1
     if [[ "$missing" ]]; then
         echo "[ERROR] docker command not found, please install docker. https://docs.docker.com/installation/" | red
-        exit 127
+        _exit 127
     fi
     info "docker command: OK"
 
     local ver=$(docker --version 2> /dev/null)
     local numver=$(docker-getversion $ver)
     
-    if [ $numver -lt 150 ]; then
+    if [ $numver -lt 180 ]; then
         local target=$(which docker 2>/dev/null || true)
         : ${target:=/usr/local/bin/docker}
-        echo "[ERROR] Please upgrade your docker version to 1.5.0 or latest" | red
+        echo "[ERROR] Please upgrade your docker version to 1.8.0 or latest" | red
         echo "suggested command:" | red
         echo "  sudo curl -Lo $target https://get.docker.com/builds/$(uname -s)/$(uname -m)/docker-latest ; chmod +x $target" | blue
-        exit 1
+        _exit 1
     fi
     info "docker client version: OK"
 }
 
 docker-check-server-version() {
-    docker version &> /tmp/cbd.log || noserver=1
+    docker version &> $TEMP_DIR/cbd.log || noserver=1
     if [[ "$noserver" ]]; then
         echo "[ERROR] docker version returned an error" | red
-        cat /tmp/cbd.log | yellow
-        exit 127
+        cat $TEMP_DIR/cbd.log | yellow
+        _exit 127
     fi
 
     local numserver
@@ -118,17 +118,23 @@ docker-check-server-version() {
     fi
 
 
-    if [ $numserver -lt 150 ]; then
-        echo "[ERROR] Please upgrade your docker version to 1.5.0 or latest" | red
+    if [ $numserver -lt 180 ]; then
+        echo "[ERROR] Please upgrade your docker version to 1.8.0 or latest" | red
         echo "[WARNING] your local docker seems to be fine, only the server version is outdated" | yellow
-        exit 1
+        _exit 1
     fi
     info "docker server version: OK"
 }
 
 docker-check-version() {
-    declare desc="Checks if docker is at least 1.5.0"
+    declare desc="Checks if docker is at least 1.8.0"
 
     docker-check-client-version
     docker-check-server-version
 }
+
+docker-kill-all-sidekicks() {
+    debug "kill all exited container labeled as: cbreak.sidekick"
+    ( docker rm -f $(docker ps -qa -f 'label=cbreak.sidekick' -f status=exited ) & ) &>/dev/null
+}
+